@@ -1,22 +1,39 @@
 {
 	"name": "bafs/parvula",
-	"description": "Simple, Markdown and Fast CMS",
-	"keywords": ["cms", "markdown"],
+	"description": "An extremely simple & flexible CMS",
+	"type": "project",
+	"keywords": ["cms", "flat files", "markdown"],
 	"license": "MIT",
-	"homepage": "https://github.com/BafS/parvula",	
+	"homepage": "https://github.com/BafS/parvula",
 	"authors": [
 		{
 			"name": "BafS",
 			"email": "fabacrans@gmail.com"
 		}
 	],
+	"support": {
+		"issues": "https://github.com/BafS/parvula/issues",
+		"source": "https://github.com/BafS/parvula"
+	},
+	"minimum-stability": "stable",
 	"require": {
-		"php": ">=5.3",
-		"michelf/php-markdown": "1.4.*"
+		"php": ">=5.4",
+		"league/plates": "3.*",
+		"erusev/parsedown-extra": "^0.7.0",
+		"nikic/fast-route": "^0.7.0",
+		"ircmaxell/password-compat": "^1.0",
+		"symfony/yaml": "^2.8"
+	},
+	"require-dev": {
+		"filp/whoops": "1.*"
 	},
 	"autoload": {
 		"psr-4": {
-			"Parvula\\": "Parvula"
+			"Parvula\\": "Parvula",
+			"Plugin\\": "plugins"
 		}
+	},
+	"config": {
+		"preferred-install": "dist"
 	}
 }
