@@ -6,7 +6,7 @@ namespace Parvula\Core;
  * Configuration wrapper for config array
  *
  * @package Parvula
- * @version 0.1.0
+ * @version 0.5.0
  * @since 0.1.0
  * @author Fabien Sa
  * @license MIT License
@@ -16,58 +16,126 @@ class Config {
 	/**
 	 * @var array
 	 */
-	private static $config = array();
+	private $config = [];
 
 	/**
 	 * Populate class with config array
+	 *
 	 * @param array $config
-	 * @return
 	 */
-	public static function populate(array $config) {
-		static::$config = $config;
+	public function __construct(array $config) {
+		$this->config = $config;
 	}
 
 	/**
 	 * Append config to Config class
+	 *
 	 * @param array $config
 	 * @return
 	 */
-	public static function append(array $config) {
-		static::$config = $config + static::$config;
+	public function append(array $config) {
+		$this->config = $config + $this->config;
 	}
 
 	/**
 	 * Get configuration value from key
+	 *
 	 * @param mixed $key
+	 * @param mixed $default optional Default if value if nothing
 	 * @return mixed Value from config
 	 */
-	public static function get($key) {
-		if(isset(static::$config[$key])) {
-			return static::$config[$key];
+	public function get($key, $default = null) {
+		$pieces = explode('.', $key);
+		$ptr = &$this->config;
+
+		foreach($pieces as $step) {
+			if (!isset($ptr[$step])) {
+				return $default;
+			}
+			$ptr = &$ptr[$step];
 		}
 
-		return;
+		return $ptr;
 	}
 
 	/**
-	 * Set configuration value from key
+	 * Set configuration value from key (create a new key if needed)
+	 *
 	 * @param mixed $key
 	 * @param mixed $value
-	 * @return
+	 * @param bool True if value is set
 	 */
-	public static function set($key, $value) {
-		if(!empty($key)) {
-			static::$config[$key] = $value;
+	public function set($key, $value) {
+		$pieces = explode('.', $key);
+		$ptr = &$this->config;
+
+		foreach($pieces as $step) {
+			$ptr = &$ptr[$step];
+			if (!isset($ptr[$step])) {
+				$ptr = [];
+			}
 		}
+
+		return $ptr = $value;
 	}
 
 	/**
-	 * Shortcut to {@see Config::get} method
+	 * Edit configuration value from key (without creating a new key)
+	 *
 	 * @param mixed $key
-	 * @return mixed Value from config
+	 * @param mixed $value
+	 * @param bool False if the value does not exists
 	 */
-	public static function __callStatic($key, $_) {
-		return static::get($key);
+	public function edit($key, $value) {
+		$pieces = explode('.', $key);
+		$ptr = &$this->config;
+
+		foreach($pieces as $step) {
+			if (!isset($ptr[$step])) {
+				return false;
+			}
+			$ptr = &$ptr[$step];
+		}
+
+		$ptr = $value;
+		return true;
+	}
+
+	/**
+	 * Check if key exists
+	 *
+	 * @return bool
+	 */
+	public function has($key) {
+		$pieces = explode('.', $key);
+		$ptr = &$this->config;
+
+		foreach($pieces as $step) {
+			if (!isset($ptr[$step])) {
+				return false;
+			}
+			$ptr = &$ptr[$step];
+		}
+
+		return true;
+	}
+
+	/**
+	 * Get the configuration as an array
+	 *
+	 * @return array
+	 */
+	public function toArray() {
+		return $this->config;
+	}
+
+	/**
+	 * Get the configuration as an object
+	 *
+	 * @return object
+	 */
+	public function toObject() {
+		return (object) $this->config;
 	}
 
 }
